Spawn More Workers When There Is Work To Do
Level 5 rooms frequently collapse. They are at a sensitive point where the code base switches from non-storage to storage base behaviour and where links are introduced.  If the player does not sequence the construction of these facilities or if one is destroyed it can cause a room to struggle or fail.  

Rooms with long haulage distances from the energy source to the spawn/extension/storage also suffer greatly as we rely heavily on haulers and workers to do the transport of energy.  If there are no working link facilities the workers and haulers can spend all their time hauling energy and recharging the spawn.  In the worst cases they do this at a rate too low to support the continued generation or workers, haulers and miners.  

This change maintains the current behaviour of always having at least a value of one for the maximum number of workers.  This still leaves us open to that worker expiring and us relying on a hauler to energise a spawn.  However if there is no hauler in the room we will get a maximum number of workers of at least 2, which provides some protection for the times when one worker fades out.  Further we get another bonus worker for any existing construction sites (such as when there is a link to finish), which seems to help complete construction in my long haul rooms.  Finally we get another bonus worker if the room is fortifiable, this is nothing to do with the room fragility but it does stick with the idea of spending energy when there is excess.  

Using this code (and the other change I am about to add to hauler counts) my failing level 5 room recovered, completed its buildings (after 24 hours of making no progress on them) and is now thriving.  I am back down to 2 workers as I have excess energy to fortify the place.