Spawn More Haulers When Drop Mining Energy
In long haul rooms energy is drop mined once the container by a distant spawn is full.  This causes workers and haulers to focus on moving energy.  The problem with this is that workers then do very little or no building work, as they spend all their time recharging distance spawns and extensions.  

To prevent this focus on charging in workers, we can spawn additional haulers whenever there is a significant (1000 in this case) amount of energy in the proximity of a spawn.  This should be enough help to give workers a chance to do some building, hopefully building of a pair of links to simplify the energy transfer and eliminate the need for the additional worker.  

With this change and the additional worker change I just introduced my struggling level 5 room was able to finally complete its building tasks (after 24 hours of having them stalled) and has begun to flourish.  Now the links are build there is no energy on the ground and I am back down the normal number of haulers.